+++ Client/src/components/Modal/ToastModal.tsx
@@ -10,7 +10,7 @@ const ToastModal = () => {
 
   return (
     <div className="w-full h-full fixed top-0 flex flex-col justify-center items-center font-['Anton-Regular'] bg-[#4a4b4a42] text-[#838282]">
-      <div className="w-[300px] flex flex-col justify-center items-center border-2 border-gray-500 border-solid animate-fadeIn">
+      <div className="w-[300px] flex flex-col justify-center items-center bg-[#414052] border-2 border-gray-500 border-solid animate-fadeIn">
         <div className="my-8 text-[#ffff]">
           <p className="mb-4">플리 제목</p>
           <input type="text" className="w-[250px] pl-2 rounded-2xl bg-[#6b6767]"></input>
