+++ Client/src/redux/store.ts
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
 import ModalSlice from './slice/ModalSlice';
+import PlaylistsSlice from './slice/PlaylistsSlice';
 
 const store = configureStore({
   reducer: {
     modal: ModalSlice.reducer,
+    playlists: PlaylistsSlice.reducer,
   },
 });
